Refactor:  move db interactions,some shape actions
